pc-6 wip fixing time out with userRole
