+++ testcafe/04_menu.js
@@ -1,6 +1,7 @@
 import { Selector } from 'testcafe'
 
 import { youngUserRole } from './helpers/roles'
+import { ROOT_PATH } from '../src/utils/config'
 
 const menuButton = Selector('#open-menu-button')
 const mainMenu = Selector('#main-menu')
@@ -8,6 +9,7 @@ const mainMenu = Selector('#main-menu')
 fixture`04_01 Menu - Affichage de la modale`.beforeEach(async t => {
   await t
     .useRole(youngUserRole)
+    .navigateTo(`${ROOT_PATH}mentions-legales`)
     .wait(500)
     .click(menuButton)
     .wait(100)
@@ -41,6 +43,7 @@ test('Je vois le montant de mon pass dans le header', async t => {
 fixture`04_02 Modale Menu - Liens vers pages`.beforeEach(async t => {
   await t
     .useRole(youngUserRole)
+    .navigateTo(`${ROOT_PATH}mentions-legales`)
     .wait(500)
     .click(menuButton)
     .wait(100)
